server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
resend==0.8.0
pydantic[email]==2.5.3
